receipts/requirements.txt
asgiref==3.7.1
astroid==2.15.0
attrs==22.2.0
autopep8==2.0.2
boto3==1.26.89
botocore==1.29.89
click==8.1.3
dependency-injector==4.41.0
dill==0.3.6
Flask==2.2.3
gunicorn==20.1.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.5.0
platformdirs==3.1.1
pluggy==1.0.0
pycodestyle==2.10.0
pylint==2.17.0
pylint-flask==0.6
pylint-plugin-utils==0.7
pytesseract==0.3.10
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.6
urllib3==1.26.15
uuid==1.30
Werkzeug==2.2.3
wrapt==1.15.0
